requirements.txt
asgiref==3.5.0
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
Django==4.0.4
django-extensions==3.1.5
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==21.3
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycparser==2.21
pyparsing==3.0.8
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
